feat: 实现飞机大战模难度分级

+++ src/edu/hitsz/aircraft/HeroAircraft.java
@@ -25,7 +25,7 @@ public class HeroAircraft extends AbstractAircraft {
      * @param speedY 英雄机射出的子弹的基准速度（英雄机无特定速度）
      * @param hp    初始生命值
      */
-    private HeroAircraft(int locationX, int locationY, int speedX, int speedY, int hp) {
+    private HeroAircraft(int locationX, int locationY, double speedX, double speedY, int hp) {
         super(locationX, locationY, speedX, speedY, hp);
         this.shootNum = 1;
         this.power = 30;
