+++ src/edu/hitsz/prop/RingBulletProp.java
@@ -4,7 +4,7 @@ import edu.hitsz.strategy.RingShootStrategy;
 
 public class RingBulletProp extends BulletProp {
 
-    public RingBulletProp(int locationX, int locationY, int speedX, int speedY) {
+    public RingBulletProp(int locationX, int locationY, double speedX, double speedY) {
         super(locationX, locationY, speedX, speedY);
         this.shootNum = 20;
         this.shootStrategy = new RingShootStrategy();
