+++ src/edu/hitsz/prop/ScatterBulletProp.java
@@ -4,7 +4,7 @@ import edu.hitsz.strategy.ScatterShootStrategy;
 
 public class ScatterBulletProp extends BulletProp {
 
-    public ScatterBulletProp(int locationX, int locationY, int speedX, int speedY) {
+    public ScatterBulletProp(int locationX, int locationY, double speedX, double speedY) {
         super(locationX, locationY, speedX, speedY);
         this.shootNum = 3;
         this.shootStrategy = new ScatterShootStrategy();
